+++ data-raw/sensor.R
@@ -1,6 +1,6 @@
 # create sensor vector
 library(rwalkr)
-sensor <- unique(walk_melb(from = Sys.Date() - 1L)$Sensor)
+sensor <- unique(walk_melb(from = Sys.Date() - 1L, tweak = TRUE)$Sensor)
 sensor_abbr <- gsub(" ", "", gsub("[:a-z:]", "", sensor))
 sensor_df <- data.frame(
   sensor = sensor, abbr = sensor_abbr, 
@@ -8,11 +8,12 @@ sensor_df <- data.frame(
 )
 
 # create a dictionary for sensor names between walk_melb and run_melb
+walk_sensor <- unique(walk_melb(from = Sys.Date() - 1L)$Sensor)
 run_sensor <- run_melb()
 run_sensor <- unique(run_sensor$Sensor)
-same_sensor <- intersect(sensor, run_sensor)
-diff_sensor <- sensor[!(sensor %in% run_sensor)]
-base_sensor <- run_sensor[!(run_sensor %in% sensor)]
+same_sensor <- intersect(walk_sensor, run_sensor)
+diff_sensor <- walk_sensor[!(walk_sensor %in% run_sensor)]
+base_sensor <- run_sensor[!(run_sensor %in% walk_sensor)]
 diff_df <- data.frame(
   "run" = c(
     "Flinders Street Station Underpass",
@@ -22,12 +23,14 @@ diff_df <- data.frame(
     "Flinders St-Spark La",
     "Queen St (West)",
     "Lygon St (East)",
-    "Lonsdale St-Spring St (West)",
+    NA,
     "St Kilda Rd-Alexandra Gardens",
     "Flinders La-Swanston St (West)",
-    "Little Collins St-Swanston St (East)"
+    "Little Collins St-Swanston St (East)",
+    "Pelham St (S)",
+    "Lonsdale St-Spring St (West)"
   ),
-  "walk" = c(diff_sensor, rep("NA", length(base_sensor) - length(diff_sensor))),
+  "walk" = c(diff_sensor, rep(NA, length(base_sensor) - length(diff_sensor) + 1)),
   "match" = FALSE,
   stringsAsFactors = FALSE
 )